pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["VladimirVlasov1982 <vladimir.vlasov82@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
pydantic = "^1.10.4"
requests = "^2.28.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
